requirements.txt
pymupdf
pdfplumber
faiss-cpu
dashscope
langgraph
langchain-core
python-dotenv
numpy
